Added x64 targets, swapped deprecated hash_map for unordered_map, fixed mouse controls Also widened field of view, added explicit static_casts for calls to .size() and added error parameter to printf for glsl_program

+++ DualContouringSample/octree.cpp
@@ -236,7 +236,7 @@ void GenerateVertexIndices(OctreeNode* node, VertexBuffer& vertexBuffer)
 			exit(EXIT_FAILURE);
 		}
 
-		d->index = vertexBuffer.size();
+		d->index = static_cast<int>(vertexBuffer.size());
 		vertexBuffer.push_back(MeshVertex(d->position, d->averageNormal));
 	}
 }
